requirements.txt
pyyaml
ollama
pandas
colorama
kagglehub
rapidfuzz